refactor template writer
