+++ pkg/osutil/osutil_test.go
@@ -6,7 +6,6 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/Azure/draft/template"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -91,15 +90,4 @@ func TestEnsureFile(t *testing.T) {
 	assert.FileExists(t, invalidFile)
 
 	os.Remove(invalidFile)
-}
-
-func TestCopyDirToFileMap(t *testing.T) {
-
-	templatewriter := &FileMapWriter{}
-	err := CopyDir(template.Dockerfiles, "dockerfiles/javascript", "/test/dir", nil, map[string]string{
-		"PORT": "8080",
-	}, templatewriter)
-	assert.Nil(t, err)
-	assert.NotNil(t, templatewriter.FileMap)
-	assert.NotNil(t, templatewriter.FileMap["/test/dir/Dockerfile"])
 }
