+++ pkg/templatewriter/templatewriter.go
@@ -1,0 +1,6 @@
+package templatewriter
+
+type TemplateWriter interface {
+	WriteFile(string, []byte) error
+	EnsureDirectory(string) error
+}
